Agregar clases de Cliente, Proveedor y Transacción, y sus respectivas pruebas a carpetas separadas
